requirements.txt
pip==23.0.1
click==8.0.4
colorama==0.4.6
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.3
pillow==9.4.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cpu
# pytorch-lightning==2.0.1
torchvision==0.15.1+cpu
torchmetrics==0.11.4
pyocr==0.8.3
pytesseract==0.3.10
easyocr==1.6.2
numpy==1.23.5
opencv-python==4.5.5.64
